+++ Assets.xcassets/AccentColor.colorset/Contents.json
@@ -2,8 +2,13 @@
   "colors" : [
     {
       "color" : {
-        "platform" : "universal",
-        "reference" : "systemCyanColor"
+        "color-space" : "display-p3",
+        "components" : {
+          "alpha" : "1.000",
+          "blue" : "0.452",
+          "green" : "0.450",
+          "red" : "0.943"
+        }
       },
       "idiom" : "universal"
     },
