+++ Sources/App/Container.swift
@@ -5,42 +5,28 @@ import AVKit
 public extension AppContainer {
     @SceneBuilder static func rootScene(store: Store) -> some SwiftUI.Scene {
         rootSceneFacet(store: store)
-        //rootSceneClassic(store: store)
     }
 
     @SceneBuilder static func rootSceneFacet(store: Store) -> some SwiftUI.Scene {
         WindowGroup { // or DocumentGroup
-            FacetHostingView(store: store).environmentObject(store)
-        }
-        .commands {
-            SidebarCommands()
-            FacetCommands(store: store)
-        }
-    }
-
-    @SceneBuilder static func rootSceneClassic(store: Store) -> some SwiftUI.Scene {
-        WindowGroup {
-            ContentView()
-                .windowToolbarUnified(compact: true, showsTitle: true)
+            FacetHostingView(store: store)
                 .environmentObject(store)
                 .task {
                     //await store.createStatusItems()
                     //await store.setDockMenu()
                     do {
-                        #if os(iOS)
+#if os(iOS)
                         try AVAudioSession.sharedInstance().setCategory(.playback)
                         try AVAudioSession.sharedInstance().setActive(true)
-                        #endif
+#endif
                     } catch {
                         dbg("error setting up session:", error)
                     }
                 }
-            // iOS only
-            // .onReceive(NotificationCenter.default.publisher(for: UXApplication.didEnterBackgroundNotification)) { _ in
-            //      dbg("didEnterBackgroundNotification")
-            //            AVAudioSession.sharedInstance
         }
         .commands {
+            SidebarCommands()
+            FacetCommands(store: store)
             CommandGroup(replacing: CommandGroupPlacement.newItem) {
                 // only permit a single window
             }
